+++ src/Moq/Protected/IProtectedMock.cs
@@ -1,6 +1,7 @@
 // Copyright (c) 2007, Clarius Consulting, Manas Technology Solutions, InSTEDD.
 // All rights reserved. Licensed under the BSD 3-Clause License; see License.txt.
 
+using System;
 using System.ComponentModel;
 using System.Diagnostics.CodeAnalysis;
 
@@ -49,6 +50,17 @@ namespace Moq.Protected
 		ISetup<TMock> Setup(string voidMethodName, bool exactParameterMatch, params object[] args);
 
 		/// <summary>
+		/// Specifies a setup for a void method invocation with the given
+		/// <paramref name="voidMethodName"/>, optionally specifying arguments for the method call.
+		/// </summary>
+		/// <param name="voidMethodName">The name of the void method to be invoked.</param>
+		/// <param name="genericTypeArguments">An array of types to be substituted for the type parameters of the current generic method definition.</param>
+		/// <param name="exactParameterMatch">Should the parameter types match exactly types that were provided</param>
+		/// <param name="args">The optional arguments for the invocation. If argument matchers are used,
+		/// remember to use <see cref="ItExpr"/> rather than <see cref="It"/>.</param>
+		ISetup<TMock> Setup(string voidMethodName, Type[] genericTypeArguments, bool exactParameterMatch, params object[] args);
+
+		/// <summary>
 		/// Specifies a setup for an invocation on a property or a non void method with the given 
 		/// <paramref name="methodOrPropertyName"/>, optionally specifying arguments for the method call.
 		/// </summary>
@@ -68,6 +80,18 @@ namespace Moq.Protected
 		/// <param name="exactParameterMatch">Should the parameter types match exactly types that were provided</param>
 		/// <typeparam name="TResult">The return type of the method or property.</typeparam>
 		ISetup<TMock, TResult> Setup<TResult>(string methodOrPropertyName, bool exactParameterMatch, params object[] args);
+
+		/// <summary>
+		/// Specifies a setup for an invocation on a property or a non void method with the given 
+		/// <paramref name="methodOrPropertyName"/>, optionally specifying arguments for the method call.
+		/// </summary>
+		/// <param name="methodOrPropertyName">The name of the method or property to be invoked.</param>
+		/// <param name="genericTypeArguments">An array of types to be substituted for the type parameters of the current generic method definition.</param>
+		/// <param name="args">The optional arguments for the invocation. If argument matchers are used, 
+		/// remember to use <see cref="ItExpr"/> rather than <see cref="It"/>.</param>
+		/// <param name="exactParameterMatch">Should the parameter types match exactly types that were provided</param>
+		/// <typeparam name="TResult">The return type of the method or property.</typeparam>
+		ISetup<TMock, TResult> Setup<TResult>(string methodOrPropertyName, Type[] genericTypeArguments, bool exactParameterMatch, params object[] args);
 
 		/// <summary>
 		/// Specifies a setup for an invocation on a property getter with the given 
@@ -105,6 +129,16 @@ namespace Moq.Protected
 		ISetupSequentialAction SetupSequence(string methodOrPropertyName, bool exactParameterMatch, params object[] args);
 
 		/// <summary>
+		/// Performs a sequence of actions, one per call.
+		/// </summary>
+		/// <param name="methodOrPropertyName">Name of the method or property being set up.</param>
+		/// <param name="genericTypeArguments">An array of types to be substituted for the type parameters of the current generic method definition.</param>
+		/// <param name="exactParameterMatch">Determines whether the parameter types should exactly match the types provided.</param>
+		/// <param name="args">The optional arguments for the invocation. If argument matchers are used,
+		/// remember to use <see cref="ItExpr"/> rather than <see cref="It"/>.</param>
+		ISetupSequentialAction SetupSequence(string methodOrPropertyName, Type[] genericTypeArguments, bool exactParameterMatch, params object[] args);
+
+		/// <summary>
 		/// Return a sequence of values, once per call.
 		/// </summary>
 		/// <param name="methodOrPropertyName">Name of the method or property being set up.</param>
@@ -122,6 +156,17 @@ namespace Moq.Protected
 		/// remember to use <see cref="ItExpr"/> rather than <see cref="It"/>.</param>
 		/// <typeparam name="TResult">Return type of the method or property being set up.</typeparam>
 		ISetupSequentialResult<TResult> SetupSequence<TResult>(string methodOrPropertyName, bool exactParameterMatch, params object[] args);
+
+		/// <summary>
+		/// Return a sequence of values, once per call.
+		/// </summary>
+		/// <param name="methodOrPropertyName">Name of the method or property being set up.</param>
+		/// <param name="genericTypeArguments">An array of types to be substituted for the type parameters of the current generic method definition.</param>
+		/// <param name="exactParameterMatch">Determines whether the parameter types should exactly match the types provided.</param>
+		/// <param name="args">The optional arguments for the invocation. If argument matchers are used,
+		/// remember to use <see cref="ItExpr"/> rather than <see cref="It"/>.</param>
+		/// <typeparam name="TResult">Return type of the method or property being set up.</typeparam>
+		ISetupSequentialResult<TResult> SetupSequence<TResult>(string methodOrPropertyName, Type[] genericTypeArguments, bool exactParameterMatch, params object[] args);
 
 		#endregion
 
@@ -148,11 +193,40 @@ namespace Moq.Protected
 		/// <exception cref="MockException">The invocation was not call the times specified by
 		/// <paramref name="times"/>.</exception>
 		/// <param name="methodName">The name of the void method to be verified.</param>
+		/// <param name="genericTypeArguments">An array of types to be substituted for the type parameters of the current generic method definition.</param>
+		/// <param name="times">The number of times a method is allowed to be called.</param>
+		/// <param name="args">The optional arguments for the invocation. If argument matchers are used, 
+		/// remember to use <see cref="ItExpr"/> rather than <see cref="It"/>.</param>
+		void Verify(string methodName, Type[] genericTypeArguments, Times times, params object[] args);
+
+		/// <summary>
+		/// Specifies a verify for a void method with the given <paramref name="methodName"/>,
+		/// optionally specifying arguments for the method call. Use in conjunction with the default
+		/// <see cref="MockBehavior.Loose"/>.
+		/// </summary>
+		/// <exception cref="MockException">The invocation was not call the times specified by
+		/// <paramref name="times"/>.</exception>
+		/// <param name="methodName">The name of the void method to be verified.</param>
 		/// <param name="times">The number of times a method is allowed to be called.</param>
 		/// <param name="exactParameterMatch">Should the parameter types match exactly types that were provided</param>
 		/// <param name="args">The optional arguments for the invocation. If argument matchers are used, 
 		/// remember to use <see cref="ItExpr"/> rather than <see cref="It"/>.</param>
 		void Verify(string methodName, Times times, bool exactParameterMatch, params object[] args);
+
+		/// <summary>
+		/// Specifies a verify for a void method with the given <paramref name="methodName"/>,
+		/// optionally specifying arguments for the method call. Use in conjunction with the default
+		/// <see cref="MockBehavior.Loose"/>.
+		/// </summary>
+		/// <exception cref="MockException">The invocation was not call the times specified by
+		/// <paramref name="times"/>.</exception>
+		/// <param name="methodName">The name of the void method to be verified.</param>
+		/// <param name="genericTypeArguments">An array of types to be substituted for the type parameters of the current generic method definition.</param>
+		/// <param name="times">The number of times a method is allowed to be called.</param>
+		/// <param name="exactParameterMatch">Should the parameter types match exactly types that were provided</param>
+		/// <param name="args">The optional arguments for the invocation. If argument matchers are used, 
+		/// remember to use <see cref="ItExpr"/> rather than <see cref="It"/>.</param>
+		void Verify(string methodName, Type[] genericTypeArguments, Times times, bool exactParameterMatch, params object[] args);
 
 		/// <summary>
 		/// Specifies a verify for an invocation on a property or a non void method with the given 
@@ -175,6 +249,21 @@ namespace Moq.Protected
 		/// <exception cref="MockException">The invocation was not call the times specified by 
 		/// <paramref name="times"/>.</exception>
 		/// <param name="methodName">The name of the method or property to be invoked.</param>
+		/// <param name="genericTypeArguments">An array of types to be substituted for the type parameters of the current generic method definition.</param>
+		/// <param name="args">The optional arguments for the invocation. If argument matchers are used, 
+		/// remember to use <see cref="ItExpr"/> rather than <see cref="It"/>.</param>
+		/// <param name="times">The number of times a method is allowed to be called.</param>
+		/// <typeparam name="TResult">The type of return value from the expression.</typeparam>
+		[SuppressMessage("Microsoft.Design", "CA1004:GenericMethodsShouldProvideTypeParameter")]
+		void Verify<TResult>(string methodName, Type[] genericTypeArguments, Times times, params object[] args);
+
+		/// <summary>
+		/// Specifies a verify for an invocation on a property or a non void method with the given 
+		/// <paramref name="methodName"/>, optionally specifying arguments for the method call.
+		/// </summary>
+		/// <exception cref="MockException">The invocation was not call the times specified by 
+		/// <paramref name="times"/>.</exception>
+		/// <param name="methodName">The name of the method or property to be invoked.</param>
 		/// <param name="exactParameterMatch">Should the parameter types match exactly types that were provided</param>
 		/// <param name="args">The optional arguments for the invocation. If argument matchers are used, 
 		/// remember to use <see cref="ItExpr"/> rather than <see cref="It"/>.</param>
@@ -182,6 +271,22 @@ namespace Moq.Protected
 		/// <typeparam name="TResult">The type of return value from the expression.</typeparam>
 		[SuppressMessage("Microsoft.Design", "CA1004:GenericMethodsShouldProvideTypeParameter")]
 		void Verify<TResult>(string methodName, Times times, bool exactParameterMatch, params object[] args);
+
+		/// <summary>
+		/// Specifies a verify for an invocation on a property or a non void method with the given 
+		/// <paramref name="methodName"/>, optionally specifying arguments for the method call.
+		/// </summary>
+		/// <exception cref="MockException">The invocation was not call the times specified by 
+		/// <paramref name="times"/>.</exception>
+		/// <param name="methodName">The name of the method or property to be invoked.</param>
+		/// <param name="genericTypeArguments">An array of types to be substituted for the type parameters of the current generic method definition.</param>
+		/// <param name="exactParameterMatch">Should the parameter types match exactly types that were provided</param>
+		/// <param name="args">The optional arguments for the invocation. If argument matchers are used, 
+		/// remember to use <see cref="ItExpr"/> rather than <see cref="It"/>.</param>
+		/// <param name="times">The number of times a method is allowed to be called.</param>
+		/// <typeparam name="TResult">The type of return value from the expression.</typeparam>
+		[SuppressMessage("Microsoft.Design", "CA1004:GenericMethodsShouldProvideTypeParameter")]
+		void Verify<TResult>(string methodName, Type[] genericTypeArguments, Times times, bool exactParameterMatch, params object[] args);
 
 		/// <summary>
 		/// Specifies a verify for an invocation on a property getter with the given